+++ cartoframes/viz/source.py
@@ -23,7 +23,7 @@ class Source(object):
 
     Args:
         data (str, :py:class:`GeoFrame <geopandas.GeoDataFrame>`,
-          :py:class:`Dataset <cartoframes.vis.Dataset>` ): a table name, SQL query
+          :py:class:`Dataset <cartoframes.viz.Dataset>` ): a table name, SQL query
           ,GeoJSON file, GeoFrame object or Dataset object.
         context (:py:class:`Context <cartoframes.Context>`):
           A Conext instance. If not provided the context will be automatically
@@ -40,7 +40,7 @@ class Source(object):
         .. code::
 
             from cartoframes import Context, set_default_context
-            from cartoframes.vis import Source
+            from cartoframes.viz import Source
 
             context = Context(
                 base_url='https://your_user_name.carto.com',
@@ -55,7 +55,7 @@ class Source(object):
         .. code::
 
             from cartoframes import Context, set_default_context
-            from cartoframes.vis import Source
+            from cartoframes.viz import Source
 
             context = Context(
                 base_url='https://your_user_name.carto.com',
@@ -70,7 +70,7 @@ class Source(object):
         .. code::
 
             from cartoframes import Context, set_default_context
-            from cartoframes.vis import Source
+            from cartoframes.viz import Source
 
             context = Context(
                 base_url='https://your_user_name.carto.com',
@@ -85,7 +85,7 @@ class Source(object):
         .. code::
 
             from cartoframes import Context, set_default_context
-            from cartoframes.vis import Source
+            from cartoframes.viz import Source
             from cartoframes import Dataset
 
             context = Context(
@@ -103,7 +103,7 @@ class Source(object):
         .. code::
 
             from cartoframes import Context
-            from cartoframes.vis import Source
+            from cartoframes.viz import Source
 
             context = Context(
                 base_url='https://your_user_name.carto.com',
@@ -117,7 +117,7 @@ class Source(object):
         .. code::
 
             from cartoframes import Context, set_default_context
-            from cartoframes.vis import Source
+            from cartoframes.viz import Source
 
             context = Context(
                 base_url='https://your_user_name.carto.com',